Added a special flag for global tags
Previously the field binding type was being appended to the id for the cid for each field. This created a collision when there was a normal field with the id of say "16" and then a global field with an id of 16 on the same page. The '2nd' field on the page would bind to the location of the first field because they would both have the same cid. I added a 'global' string in-between so now it's 'fieldglobal16'