Enabler added
Slices now enabled by default and do not show the enable property controls unless "enabler" property is set on the config. Additionally, we do a better job of shoving in all the properties of a configuration into the field for further customization. We are using it for this new "enabler" option